+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Update django to 3.2.11
 - New page: list past and future interviews
 - Fix an error when we reopen a closed process wthout interview
+- Show kind of interview on plan form and minute form
 
 # v1.13.0 (2022-01-26)
 - Add interview.prequalification to interviews TSV export