+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.14.0
-- Update django to 3.2.11
+- Update django to 3.2.12
 - New page: list past and future interviews
 - Fix an error when we reopen a closed process wthout interview
 - Show kind of interview on plan form and minute form